requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth==2.23.4
requests==2.31.0
Pillow==11.0.0
python-dotenv==1.0.1
httplib2==0.22.0
